refactor: move get_verification_message to lib.rs
