+++ tee-worker/omni-executor/heima/identity-verification/src/web3.rs
@@ -1,23 +1,4 @@
-use executor_crypto::hashing::blake2_256;
 use executor_primitives::{utils::hex::hex_encode, Identity, Web3ValidationData};
-use parity_scale_codec::Encode;
-
-// verification message format:
-// ```
-// blake2_256(<parentchain nonce> + <sender identity> + <identity-to-be-added>)
-// ```
-// where <> means SCALE-encoded
-pub fn get_verification_message(
-	member: &Identity,
-	identity_to_add: &Identity,
-	omni_account_nonce: u64,
-) -> Vec<u8> {
-	let mut payload = Vec::new();
-	payload.append(&mut omni_account_nonce.encode());
-	payload.append(&mut member.encode());
-	payload.append(&mut identity_to_add.encode());
-	blake2_256(payload.as_slice()).to_vec()
-}
 
 // This function validates the signature with both the raw message and its prettified format. Any of them is valid.
 // The prettified version was introduced to extend the support for utf-8 signatures for browser wallets that
