+++ tee-worker/omni-executor/native-task-handler/src/lib.rs
@@ -8,12 +8,10 @@ use executor_crypto::{
 	aes256::{aes_encrypt_default, Aes256Key},
 	jwt,
 };
-use executor_primitives::{
-	intent::Intent, Identity, MemberAccount, OmniAccountAuthType, ValidationData,
-};
+use executor_primitives::{intent::Intent, MemberAccount, OmniAccountAuthType, ValidationData};
 use executor_storage::{MemberOmniAccountStorage, Storage, StorageDB};
 use heima_authentication::auth_token::AuthTokenClaims;
-use heima_identity_verification::{web2, web3};
+use heima_identity_verification::{get_verification_message, web2, web3};
 use parentchain_api_interface::runtime_types::{
 	frame_system::pallet::Call as SystemCall, pallet_balances::pallet::Call as BalancesCall,
 	pallet_omni_account::pallet::Call as OmniAccountCall, paseo_parachain_runtime::RuntimeCall,
@@ -271,8 +269,7 @@ async fn handle_native_task<
 				}
 				return;
 			};
-			let verification_message =
-				web3::get_verification_message(&sender_identity, &identity, nonce);
+			let verification_message = get_verification_message(&sender_identity, &identity, nonce);
 
 			let validation_result = match validation_data {
 				ValidationData::Web2(web2_validation_data) => {
